changed indentation to 2 spaces, removed suffix DELETE_ME
