+++ contracts/strategies/sushi/SushiStrategyULMainnet_SUSHI_WETH.sol
@@ -1,9 +1,9 @@
 pragma solidity 0.5.16;
 
-import "../../base/sushi-base/MasterChefV2StrategyUL.sol";
+import "../../base/sushi-base/MasterChefStrategyUL.sol";
 
-contract SushiV2StrategyULMainnet_WETH_PSP_DELETE_ME is MasterChefV2StrategyUL {
-  address private sushi_spell_weth; // just a differentiator for the bytecode
+contract SushiStrategyULMainnet_SUSHI_WETH is MasterChefStrategyUL {
+  address public sushi_weth_unused; // just a differentiator for the bytecode
 
   constructor() public {}
 
@@ -13,36 +13,32 @@ contract SushiV2StrategyULMainnet_WETH_PSP_DELETE_ME is MasterChefV2StrategyUL {
   {
     // SushiSwap ETH/PSP LP (SLP) https://etherscan.io/address/0x458ae80894a0924ac763c034977e330c565f1687
     address underlying_ = address(
-      0x458ae80894A0924Ac763C034977e330c565F1687
+      0x795065dCc9f64b5614C407a6EFDC400DA6221FB0
     );
     address sushi = address(0x6B3595068778DD592e39A122f4f5a5cF09C90fE2);
     address weth = address(0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2);
-    address psp = address(0xcAfE001067cDEF266AfB7Eb5A286dCFD277f3dE5);
 
-    address masterchefV2 = address(
-      0xEF0881eC094552b2e128Cf945EF17a6752B4Ec5d
+    address masterchefV1 = address(
+      0xc2EdaD668740f1aA35E4D8f227fB8E17dcA888Cd
     );
     address sushiswapRouterV2 = address(
       0xd9e1cE17f2641f24aE83637ab66a2cca9C378B9F
     );
 
-    uint256 poolId_ = 31;
+    uint256 poolId_ = 12;
     bytes32 sushiDex = bytes32(uint256(keccak256("sushi")));
 
-    MasterChefV2StrategyUL.initializeBaseStrategy({
+    MasterChefStrategyUL.initializeBaseStrategy({
       __storage: __storage,
       _underlying: underlying_,
       _vault: _vault,
-      _rewardPool: masterchefV2,
+      _rewardPool: masterchefV1,
       _rewardToken: sushi,
       _poolID: poolId_,
       _routerV2: sushiswapRouterV2
     });
 
-    //sell rewardToken(=SUSHI) for WETH and PSP to be able to add liquidity to the pool
-    storedLiquidationPaths[sushi][psp] = [sushi, weth, psp];
-    storedLiquidationDexes[sushi][psp] = [sushiDex, sushiDex];
-
+    //sell rewardToken(=SUSHI) for WETH
     storedLiquidationPaths[sushi][weth] = [sushi, weth];
     storedLiquidationDexes[sushi][weth] = [sushiDex];
   }
